--- next.config.js
@@ -1,20 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  eslint: {
-    dirs: ['src']
-  },
-  reactStrictMode: true,
-  swcMinify: true,
-  experimental: {
-    optimizeCss: true,
-    appDir: false
-  }
-}
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const withPWA = require('next-pwa')({
-  dest: 'public',
-  register: true,
-  skipWaiting: true
-})
-
-module.exports = withPWA(nextConfig)
